+++ Changes.md
@@ -1,6 +1,11 @@
 0.61.x.x (relative to 0.61.1.1)
 ========
 
+Improvements
+-----
+
+- ArnoldOptions : Added `ai:dielectric_priorities`.
+
 API
 ---